+++ src/main/java/org/openlmis/stockmanagement/i18n/MessageKeys.java
@@ -244,6 +244,9 @@ public abstract class MessageKeys {
       PHYSICAL_INVENTORY_ERROR_PREFIX + ".draft.exists";
   public static final String ERROR_PHYSICAL_INVENTORY_DRAFT_SUBMIT =
       PHYSICAL_INVENTORY_ERROR_PREFIX + ".draft.submit";
+  // soh
+  private static final String SOH_ERROR_PREFIX = ERROR_PREFIX + ".stockOnHand";
+  public static final String ERROR_SOH_NOT_FOUND = SOH_ERROR_PREFIX + ".notFound";
 
   public static final String ERROR_SIZE_NULL
           = ERROR_PREFIX + ".pageable.size.null";